Building CollegiateWordRatio Calculator
Using the Apache POI library, the CollegiateWordRatio class creates two
sets of words. The first is the list of SAT words, the second is the
words from the document specified from the spreadsheet column I am
currently working with. TODO - have this application write into the
spreadsheet the word count, the collegiate word number, and the ratio.
After I have a solid sample, I will begin creating the bell curve for
comparison.